locations: move LocationForm to Views/
Mechanical move per #4331:

- `Components::LocationForm` → `Views::Controllers::Locations::Form`
- File: `app/components/location_form.rb` →
  `app/views/controllers/locations/form.rb`
- Test: `test/components/location_form_test.rb` →
  `test/views/controllers/locations/form_test.rb`
- Updated render callers in `app/views/controllers/locations/
  {new,edit}.html.erb`.

Form id is auto-derived now (drops the explicit `id: "location_form"`
in the super call) — the controller-segment fallback in
`ApplicationForm#derive_form_id` produces `"location_form"` from the
namespace.

Cleanup-in-scope (per testing rule):
- Fixed a forbidden `assert_html(html, "body", text: ...)` in the
  locked-display test by scoping to `.help-block` (the actual
  element that carries the `show_location_locked` text).

164 tests pass, lint clean.
