+++ app/views/controllers/locations/edit.html.erb
@@ -5,7 +5,7 @@ add_context_nav(location_form_edit_tabs(location: @location))
 container_class(:full)
 %>
 
-<%= render(Components::LocationForm.new(
+<%= render(Views::Controllers::Locations::Form.new(
       @location,
       display_name: @display_name,
       dubious_where_reasons: @dubious_where_reasons
